Created a Linux Edition of keyfinder.
It is being created from scratch using Lazarus.  More to come for that.

Eventually I will merge the Linux Edition and Windows Edition into a single Lazarus Project (hopefully if there is enough API support).  I would like to try to get away from the proprietary IDEs to make getting more developers easier.  This would simplify the development process for other platforms if I could have one version for all.  The only drawback is large executable size (~10-15MB).

That being said it isn't a definite yet because of the executable size.  If anyone knows Lazarus and how to slim down the executable size then that would be a big help.

git-svn-id: svn://svn.code.sf.net/p/ekeyfinder/code/trunk@44 b14f650f-2e60-4330-827c-04cc947f8aee