Updated keyfinder.cfg to 2.1.0.06 It's not in a standard format but Product keys used to install the software is now uniformly called Key so we're a step closer.  Key and Serial are no longer interchangeable.  If there is an entry which has Serial then it's an actual serial number for the product and no the key used to install it.
Commented out a bunch of bogus or crappy entries from keyfinder.cfg.

Added a python script to assist with error checking the keyfinder.cfg when I was correcting everything.

Updated the help documentation and the script for generating htmloutput.  Apparently Python 2.7 automatically converts \n to be the Windows EOL.  That is not the case with Python 2.6 I don't believe.  Didn't really test too much on that theory.  Anyway fixed that.

git-svn-id: svn://svn.code.sf.net/p/ekeyfinder/code/trunk@103 b14f650f-2e60-4330-827c-04cc947f8aee